# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: ProxyLoom P0 API
  version: 1.0.0
  description: >-
    P0 contract for a single configured workspace and trusted administrators.
    Five administrator authentication operations are implemented. Other operations
    remain contract-only; declarations do not assert kernel capability or verification.
    Management uses the session Cookie; internal operations exist exclusively on
    a separate mTLS listener; subscription retrieval uses its mandatory path token
    independently of either credential. Scope and resource IDs are server assigned.
    JSON decoding rejects unknown fields, duplicate keys, trailing documents,
    invalid Unicode, and undocumented nulls. Errors never echo input values.
    All references in this document are local; schema validation must not fetch URLs.
servers:
  - url: /
    description: Management and public subscription listener; internal paths use their own listener.
security:
  - SessionCookie: []
...
components:
  securitySchemes:
    SessionCookie:
      type: "apiKey"
      in: "cookie"
      name: "proxyloom_session"
      description: "Opaque server-side administrator session; Secure, HttpOnly, SameSite=Lax. State-changing management operations additionally enforce Origin and X-CSRF-Token. Setup/login require Origin and JSON but no session CSRF token. Secure may be omitted only for explicit development HTTP loopback PUBLIC_URL. Never accepted as subscription or Runner authorization."
    RunnerMTLS:
      type: "mutualTLS"
      description: "Registered client certificate verified only on the dedicated internal listener. Certificate identity and server registration determine eligible builds, task types and slots. No management Cookie or subscription token is accepted."
  headers:
    RequestID:
      description: "Server-generated safe request correlation ID."
      schema: { $ref: "#/components/schemas/RequestID" }
    ETag:
      description: "Strong resource revision tag \"r<N>\". Exactly one tag, no wildcard, weak tag, list, sign, whitespace or leading zero. N fits positive int64."
      schema:
        type: "string"
        pattern: "^\"r(?:[1-9][0-9]{0,17}|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6]|9223372036854775807)\"$"
    NoStore:
      schema:
        type: "string"
        const: "private, no-store"
    RetryAfter:
      description: "Bounded retry delay in seconds."
      schema:
        type: "integer"
        minimum: 1
        maximum: 86400
  parameters:
...
    ID:
      name: "id"
      in: "path"
      required: true
      schema: { $ref: "#/components/schemas/UUID" }
    IfMatch:
      name: "If-Match"
      in: "header"
      required: true
      description: "Exactly one strong tag \"r<N>\" for the resource being changed. Missing => 428; malformed, weak, wildcard, multi-tag or overflow => 400; valid stale revision => 412. Business/generation/lease/idempotency conflicts => 409. Creates and authentication operations have no resource precondition."
      schema:
        type: "string"
        pattern: "^\"r(?:[1-9][0-9]{0,17}|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6]|9223372036854775807)\"$"
...
    Limit:
      name: "limit"
      in: "query"
      schema:
        type: "integer"
        minimum: 1
        maximum: 200
        default: 50
      description: "Bounded page size. Stable ascending (created_at, id) order unless a collection documents its immutable equivalent."
    Cursor:
      name: "cursor"
      in: "query"
      schema:
        type: "string"
        minLength: 1
        maxLength: 2048
      description: "Opaque authenticated continuation tied to scope, collection, order and normalized filter digest. A changed filter or invalid cursor is 400; authorization is checked separately on every request."
...
    LastEventID:
      name: "Last-Event-ID"
      in: "header"
      required: false
      schema: { $ref: "#/components/schemas/Counter" }
      description: "One canonical decimal sequence from the prior SSE id. Zero starts at the beginning. The server replays greater seq values; an expired retention window emits snapshot_reset and requires GET /jobs/{id}."
  responses:
    BadRequest:
      description: "Malformed JSON, duplicate/unknown field, unsafe shape, invalid path/query/header, invalid Unicode or cursor. No submitted key or value is echoed."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: "Required management session or registered internal identity is missing, invalid or expired; surfaces are not interchangeable."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: "Authenticated actor lacks permission, recent authentication, a valid current-actor confirmation or required same-origin protection."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: "Resource does not exist in the authorized scope, or existence must not be disclosed."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: "Business state, generation, idempotency or lease conflict, including COMPILE_OBSOLETE, IDEMPOTENCY_CONFLICT and LEASE_LOST. A stale resource revision is 412 instead."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    PreconditionFailed:
      description: "A syntactically valid If-Match names an outdated resource revision; no mutation was applied."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    TooLarge:
      description: "Input, decompressed data, item count or event size exceeds the configured bound."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    UnprocessableEntity:
      description: "Shape is understood but merged domain semantics are invalid, including required secret clear, unsupported capability, invalid references, excluded dependencies and DNS cycles."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    PreconditionRequired:
      description: "The versioned resource operation requires If-Match but it was omitted."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    RateLimited:
      description: "Rate, slot, concurrent execution or reserved budget limit exceeded."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unavailable:
      description: "Primary database, execution capability or safe publication is unavailable. Fail closed; never serve cached authorization or substitute a direct/empty configuration."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    JobListResponse:
      description: "Typed contract response. This operation is not registered by the foundation contract implementation."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/JobListResponse" }
...
    RunnerJobEventResponse:
      description: "Typed contract response. This operation is not registered by the foundation contract implementation."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/RunnerJobEventResponse" }
...
    JobOrBatchResponse:
      description: "Persistent child JobResponse or aggregate parent TestBatchResponse, selected by the UUID kind."
      headers:
        ETag: { $ref: "#/components/headers/ETag" }
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema:
            oneOf:
              - { $ref: "#/components/schemas/JobResponse" }
              - { $ref: "#/components/schemas/TestBatchResponse" }
...
  schemas:
    UUID:
      type: "string"
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
    RequestID:
      type: "string"
      pattern: "^[A-Za-z0-9_-]{1,64}$"
    Revision:
      type: "string"
      pattern: "^(?:[1-9][0-9]{0,17}|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6]|9223372036854775807)$"
      description: "Canonical positive decimal int64 string, 1..9223372036854775807. String encoding preserves exact values in browser clients. No signs or leading zeros."
    Counter:
      type: "string"
      pattern: "^(?:0|[1-9][0-9]{0,17}|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6]|9223372036854775807)$"
      description: "Canonical nonnegative decimal int64 string, 0..9223372036854775807. No signs or leading zeros."
...
    ErrorDetail:
      type: "object"
      additionalProperties: false
      required: []
      properties:
        field_path: { $ref: "#/components/schemas/SafeFieldPath" }
        resource_id: { $ref: "#/components/schemas/UUID" }
      minProperties: 1
    SafeError:
      type: "object"
      additionalProperties: false
      required: ["code", "message", "details"]
      properties:
        code: { $ref: "#/components/schemas/ErrorCode" }
        message:
          type: "string"
          minLength: 1
          maxLength: 512
          description: "Bounded fixed safe explanation; no input values, SQL, ciphertext, URLs, raw schema messages or core stdout/stderr."
        details:
          type: "array"
          minItems: 0
          maxItems: 64
          items: { $ref: "#/components/schemas/ErrorDetail" }
    ErrorResponse:
      type: "object"
      additionalProperties: false
      required: ["request_id", "error"]
      properties:
        request_id: { $ref: "#/components/schemas/RequestID" }
        error: { $ref: "#/components/schemas/SafeError" }
...
    PageInfo:
      type: "object"
      additionalProperties: false
      required: ["limit"]
      properties:
        limit:
          type: "integer"
          minimum: 1
          maximum: 200
          default: 50
        next_cursor:
          type: "string"
          minLength: 1
          maxLength: 2048
          description: "Opaque authenticated cursor; omitted at end. Bound to scope, collection, ordering and normalized filter digest. Cursor is not authorization."
...
    ReasonRequest:
      type: "object"
      additionalProperties: false
      required: ["reason"]
      properties:
        reason:
          type: "string"
          minLength: 1
          maxLength: 200
          description: "Administrative reason; exclude credentials and configuration."
...
    JobType:
      type: "string"
      enum: ["source_refresh", "import_parse", "compile", "config_validate", "connectivity", "download_throughput"]
    JobState:
      type: "string"
      enum: ["queued", "leased", "running", "succeeded", "failed", "canceled", "timed_out"]
...
    Verdict:
      type: "string"
      enum: ["pass", "fail", "inconclusive"]
...
    JobResponse:
      type: "object"
      additionalProperties: false
      required: ["request_id", "data"]
      properties:
        request_id: { $ref: "#/components/schemas/RequestID" }
        data: { $ref: "#/components/schemas/Job" }
    JobListResponse:
      type: "object"
      additionalProperties: false
      required: ["request_id", "data", "page"]
      properties:
        request_id: { $ref: "#/components/schemas/RequestID" }
        data:
          type: "array"
          minItems: 0
          maxItems: 200
          items: { $ref: "#/components/schemas/Job" }
        page: { $ref: "#/components/schemas/PageInfo" }
...
    TestBatchResponse:
      type: "object"
      additionalProperties: false
      required: ["request_id", "data"]
      properties:
        request_id: { $ref: "#/components/schemas/RequestID" }
        data: { $ref: "#/components/schemas/TestBatch" }
...
    JobPhase:
      type: "string"
      enum: ["queued", "leased", "fetching", "parsing", "compiling", "preparing", "validating", "starting", "running", "probing", "downloading", "collecting", "settling", "canceling", "completed"]
    RunnerPhase:
      type: "string"
      enum: ["leased", "preparing", "validating", "starting", "running", "probing", "downloading", "collecting", "settling", "canceling", "completed"]
...
    RunnerJobEventRequest:
      type: "object"
      additionalProperties: false
      required: ["job_id", "attempt", "lease_seq", "event_id", "phase", "completed", "total"]
      properties:
        job_id: { $ref: "#/components/schemas/UUID" }
        attempt:
          type: "integer"
          minimum: 1
          maximum: 100
        lease_seq: { $ref: "#/components/schemas/Revision" }
        event_id: { $ref: "#/components/schemas/UUID" }
        phase: { $ref: "#/components/schemas/RunnerPhase" }
        completed:
          type: "integer"
          minimum: 0
          maximum: 2147483647
        total:
          type: "integer"
          minimum: 0
          maximum: 2147483647
        verdict: { $ref: "#/components/schemas/Verdict" }
        error: { $ref: "#/components/schemas/SafeError" }
      description: "completed<=total is also checked by the service. Error summaries are fixed safe messages, not arbitrary Runner output. event_id deduplicates delivery inside the valid attempt."
...
    JobEvent:
      type: "object"
      additionalProperties: false
      required: ["job_id", "seq", "phase", "completed", "total"]
      properties:
        job_id: { $ref: "#/components/schemas/UUID" }
        seq: { $ref: "#/components/schemas/Revision" }
        phase: { $ref: "#/components/schemas/JobPhase" }
        completed:
          type: "integer"
          minimum: 0
          maximum: 2147483647
        total:
          type: "integer"
          minimum: 0
          maximum: 2147483647
        verdict: { $ref: "#/components/schemas/Verdict" }
        error: { $ref: "#/components/schemas/SafeError" }
      description: "SSE event=job_event; SSE id is decimal seq. Sequence is server-assigned, persisted and monotonically increasing across attempts. completed<=total. No raw stdout/stderr, arbitrary JSON data or unbounded log text."
    SnapshotResetEvent:
      type: "object"
      additionalProperties: false
      required: ["job_id", "latest_seq", "snapshot_url"]
      properties:
        job_id: { $ref: "#/components/schemas/UUID" }
        latest_seq: { $ref: "#/components/schemas/Counter" }
        snapshot_url:
          type: "string"
          pattern: "^/api/v1/jobs/[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
      description: "SSE event=snapshot_reset. Requested history is outside the retained window; close the stream and reread this same job's authorized snapshot. snapshot_url UUID must equal job_id."
...
    RunnerJobEventReceipt:
      type: "object"
      additionalProperties: false
      required: ["job_id", "attempt", "lease_seq", "event_id", "seq", "replayed"]
      properties:
        job_id: { $ref: "#/components/schemas/UUID" }
        attempt:
          type: "integer"
          minimum: 1
          maximum: 100
        lease_seq: { $ref: "#/components/schemas/Revision" }
        event_id: { $ref: "#/components/schemas/UUID" }
        seq: { $ref: "#/components/schemas/Revision" }
        replayed:
          type: "boolean"
    RunnerJobEventResponse:
      type: "object"
      additionalProperties: false
      required: ["request_id", "data"]
      properties:
        request_id: { $ref: "#/components/schemas/RequestID" }
        data: { $ref: "#/components/schemas/RunnerJobEventReceipt" }
...
paths:
...
  /api/v1/jobs:
    get:
      operationId: listJobs
      x-implementation-status: implemented
      summary: List persistent jobs with execution state distinct from verdict
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - name: state
          in: query
          schema: {$ref: '#/components/schemas/JobState'}
        - name: type
          in: query
          schema: {$ref: '#/components/schemas/JobType'}
        - name: executor
          in: query
          schema: {type: string, enum: [api_worker, runner]}
        - name: batch_id
          in: query
          schema: {$ref: '#/components/schemas/UUID'}
      responses:
        '200': {$ref: '#/components/responses/JobListResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/jobs/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      operationId: getJob
      x-implementation-status: implemented
      summary: Read a child job or parent test-batch snapshot
      description: Passing a batch_id returns TestBatchResponse with aggregate progress and all bounded child summaries. Passing a child job_id returns JobResponse. Successful execution with a failing tested node is state=succeeded and verdict=fail.
      responses:
        '200': {$ref: '#/components/responses/JobOrBatchResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/jobs/{id}/cancel:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post:
      operationId: cancelJob
      x-implementation-status: implemented
      summary: Persist cancellation of a child job or all unfinished children of a batch
      description: id may be a job_id or batch_id. Returns the updated snapshot; acknowledgement means cancellation requested, not that a running process has already exited. Budget settlement still applies to canceled attempts.
      parameters: [{$ref: '#/components/parameters/IfMatch'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReasonRequest'}
      responses:
        '202': {$ref: '#/components/responses/JobOrBatchResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/jobs/{id}/events:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      operationId: streamJobEvents
      x-implementation-status: implemented
      summary: Replay bounded safe job events over SSE
      description: >-
        Each job has stable strictly increasing seq values; SSE id is that decimal
        seq. Last-Event-ID replays later events without rerunning work. Retention
        gaps produce event=snapshot_reset carrying SnapshotResetEvent, then close;
        clients reread GET /jobs/{id}. A future sequence or malformed ID is 400.
        event=job_event carries JobEvent. Neither event contains raw stdout,
        stderr, native configuration or arbitrary event data. SSE is not the
        persistent job fact source. Heartbeat comments carry no application data.
      parameters: [{$ref: '#/components/parameters/LastEventID'}]
      responses:
        '200':
          description: Authenticated event stream; typed JSON data is framed by SSE event and id fields.
          headers:
            Cache-Control: {schema: {type: string, const: 'private, no-store'}}
          content:
            text/event-stream:
              schema: {type: string}
              x-event-schemas:
                job_event: {$ref: '#/components/schemas/JobEvent'}
                snapshot_reset: {$ref: '#/components/schemas/SnapshotResetEvent'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /internal/v1/jobs/{id}/events:
    servers: [{url: 'https://runner-api.proxyloom.invalid', description: Dedicated internal mTLS listener.}]
    parameters: [{$ref: '#/components/parameters/ID'}]
    post:
      operationId: submitRunnerJobEvent
      x-implementation-status: implemented
      x-authentication-surface: runner-mtls
      security: [{RunnerMTLS: []}]
      summary: Submit a bounded typed event under the current fenced lease
      description: Body job_id must equal path id. event_id deduplicates retries within this attempt; the server assigns stable job-wide seq. Unknown fields, raw stdout and unredacted messages are rejected, and only safe diagnostic summaries are retained.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RunnerJobEventRequest'}
      responses:
        '200': {$ref: '#/components/responses/RunnerJobEventResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '413': {$ref: '#/components/responses/TooLarge'}
        '422': {$ref: '#/components/responses/UnprocessableEntity'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
...